feat: [ui-api-module-cleanup-recovery-001] - Session-factory import activation module
