+++ ui/src/api/named-factory/api.ts
@@ -1,28 +1,28 @@
-import type { components } from "../generated/openapi";
 import {
   CurrentFactoryDefinitionError,
-  getCurrentFactoryDocument,
-  saveCurrentFactoryDocument,
-  saveFactoryForSessionDocument,
   type CurrentFactoryDocument,
+  getCurrentFactoryDocument,
 } from "../current-factory-definition";
+import { SessionFactoryAPIError } from "../session-factory/errors";
 import {
-  listFactorySessions,
-  openFactorySession,
-} from "../factory-sessions";
-import {
-  extractNamedFactoryNamesFromSessionTargets,
-} from "./import-save-mode";
+  type ActivateImportedFactoryForSessionOptions,
+  activateImportedFactoryForSession as activateImportedFactoryForSessionImpl,
+  type DiscoverSessionNamedFactoryNamesOptions,
+  discoverSessionNamedFactoryNames as discoverSessionNamedFactoryNamesImpl,
+  type FactoryImportSaveChoice,
+  type ImportFactoryValue,
+} from "../session-factory/import-activation";
 import { extractAPIErrorPayload } from "../transport";
+
 export {
   allocateFirstFreeSuffixedFactoryName,
   extractNamedFactoryNamesFromSessionTargets,
   resolveImportCreateFactoryName,
-} from "./import-save-mode";
+} from "../session-factory/import-save-mode";
 
-export type FactoryValue = components["schemas"]["Factory"];
+export type FactoryValue = ImportFactoryValue;
 
-export type FactoryImportSaveChoice = "replace_current" | "create_new_named";
+export type { FactoryImportSaveChoice };
 
 export type NamedFactoryAPIErrorCode =
   | "BAD_REQUEST"
@@ -47,18 +47,10 @@ export interface GetCurrentFactoryOptions {
   sessionID?: string | null;
 }
 
-export interface ActivateImportedFactoryForSessionOptions {
-  choice?: FactoryImportSaveChoice;
-  createFactoryName?: string;
-  existingFactoryNames?: readonly string[];
-  fetch?: typeof globalThis.fetch;
-  sessionID?: string | null;
-}
-
-export interface DiscoverSessionNamedFactoryNamesOptions {
-  fetch?: typeof globalThis.fetch;
-  sessionID?: string | null;
-}
+export type {
+  ActivateImportedFactoryForSessionOptions,
+  DiscoverSessionNamedFactoryNamesOptions,
+};
 
 export class NamedFactoryAPIError extends Error {
   public readonly code: NamedFactoryAPIErrorCode;
@@ -82,9 +74,12 @@ export async function getCurrentFactory(
   const fetchImplementation = options.fetch ?? globalThis.fetch;
 
   if (typeof fetchImplementation !== "function") {
-    throw new NamedFactoryAPIError("Current factory export is unavailable in this environment.", {
-      code: "NETWORK_ERROR",
-    });
+    throw new NamedFactoryAPIError(
+      "Current factory export is unavailable in this environment.",
+      {
+        code: "NETWORK_ERROR",
+      },
+    );
   }
 
   try {
@@ -101,129 +96,37 @@ export async function getCurrentFactory(
 export async function discoverSessionNamedFactoryNames(
   options: DiscoverSessionNamedFactoryNamesOptions = {},
 ): Promise<string[]> {
-  const fetchImplementation = options.fetch ?? globalThis.fetch;
-  const sessions = await listFactorySessions({ fetch: fetchImplementation });
-  const session = sessions.find((entry) => entry.id === normalizeSessionID(options.sessionID));
-  if (!session?.folderPath) {
-    return [];
+  try {
+    return await discoverSessionNamedFactoryNamesImpl(options);
+  } catch (error) {
+    throw toNamedFactoryAPIErrorFromSessionFactory(error);
   }
-
-  const response = await openFactorySession(
-    {
-      folderPath: session.folderPath,
-      validateOnly: true,
-    },
-    { fetch: fetchImplementation },
-  );
-
-  return extractNamedFactoryNamesFromSessionTargets(response.targets);
 }
 
 export async function activateImportedFactoryForSession(
   importedFactory: FactoryValue,
   options: ActivateImportedFactoryForSessionOptions = {},
 ): Promise<FactoryValue> {
-  if (options.choice === "create_new_named") {
-    return activateImportedFactoryCreateNamedForSession(importedFactory, options);
-  }
-
-  return activateImportedFactoryReplaceCurrentForSession(importedFactory, options);
-}
-
-async function activateImportedFactoryReplaceCurrentForSession(
-  importedFactory: FactoryValue,
-  options: ActivateImportedFactoryForSessionOptions,
-): Promise<FactoryValue> {
-  let currentDocument: CurrentFactoryDocument;
   try {
-    currentDocument = await getCurrentFactoryDocument({
-      fetch: options.fetch,
-      sessionID: options.sessionID,
-    });
-  } catch (error) {
-    throw toNamedFactoryAPIErrorFromCurrentFactoryDefinition(error);
-  }
-
-  let savedDocument: CurrentFactoryDocument;
-  try {
-    savedDocument = await saveCurrentFactoryDocument(
-      {
-        baseVersion: currentDocument.version,
-        factoryDefinition: toImportedFactoryDefinition(
-          importedFactory,
-          currentDocument.name,
-        ),
-      },
-      {
-        fetch: options.fetch,
-        sessionID: options.sessionID,
-      },
+    return await activateImportedFactoryForSessionImpl(
+      importedFactory,
+      options,
     );
   } catch (error) {
-    throw toNamedFactoryAPIErrorFromCurrentFactoryDefinition(error);
+    throw toNamedFactoryAPIErrorFromSessionFactory(error);
   }
-
-  return toActivatedFactoryValue(savedDocument);
 }
 
-async function activateImportedFactoryCreateNamedForSession(
-  importedFactory: FactoryValue,
-  options: ActivateImportedFactoryForSessionOptions,
-): Promise<FactoryValue> {
-  const createFactoryName = options.createFactoryName?.trim();
-  if (!createFactoryName) {
-    throw new NamedFactoryAPIError("A factory name is required to create a new named factory.", {
-      code: "INVALID_FACTORY_NAME",
-    });
-  }
-
-  let currentDocument: CurrentFactoryDocument;
-  try {
-    currentDocument = await getCurrentFactoryDocument({
-      fetch: options.fetch,
-      sessionID: options.sessionID,
-    });
-  } catch (error) {
-    throw toNamedFactoryAPIErrorFromCurrentFactoryDefinition(error);
-  }
-
-  const { version: _version, ...importedWithoutVersion } = importedFactory;
-  const includeVersion = shouldIncludeVersionForImportCreateNamed(
-    createFactoryName,
-    currentDocument,
-    options.existingFactoryNames,
-  );
-
-  let savedDocument: CurrentFactoryDocument;
-  try {
-    savedDocument = await saveFactoryForSessionDocument(
-      {
-        baseVersion: includeVersion ? currentDocument.version : undefined,
-        factoryDefinition: {
-          ...importedWithoutVersion,
-          name: createFactoryName,
-        },
-        includeVersion,
-        mode: "UPSERT_NAMED_AND_ACTIVATE",
-      },
-      {
-        fetch: options.fetch,
-        sessionID: options.sessionID,
-      },
-    );
-  } catch (error) {
-    throw toNamedFactoryAPIErrorFromCurrentFactoryDefinition(error);
-  }
-
-  return toActivatedFactoryValue(savedDocument);
+function toActivatedFactoryValue(
+  document: CurrentFactoryDocument,
+): FactoryValue {
+  const { version: _version, ...factoryValue } = document;
+  return factoryValue;
 }
 
-function normalizeSessionID(sessionID: string | null | undefined): string {
-  const trimmed = sessionID?.trim();
-  return trimmed ? trimmed : "~default";
-}
-
-function normalizeNamedFactoryAPIErrorCode(code: string | undefined): NamedFactoryAPIErrorCode {
+function normalizeNamedFactoryAPIErrorCode(
+  code: string | undefined,
+): NamedFactoryAPIErrorCode {
   switch (code) {
     case "BAD_REQUEST":
     case "FACTORY_ALREADY_EXISTS":
@@ -240,48 +143,6 @@ function normalizeNamedFactoryAPIErrorCode(code: string | undefined): NamedFacto
   }
 }
 
-function toImportedFactoryDefinition(
-  importedFactory: FactoryValue,
-  sessionFactoryName: string,
-): FactoryValue {
-  const { version: _version, ...importedWithoutVersion } = importedFactory;
-  return {
-    ...importedWithoutVersion,
-    name: sessionFactoryName,
-  };
-}
-
-function shouldIncludeVersionForImportCreateNamed(
-  createFactoryName: string,
-  currentDocument: CurrentFactoryDocument,
-  existingFactoryNames: readonly string[] | undefined,
-): boolean {
-  const normalizedCreateFactoryName = createFactoryName.trim();
-  if (normalizedCreateFactoryName.length === 0) {
-    return false;
-  }
-
-  if (normalizedCreateFactoryName === currentDocument.name.trim()) {
-    return true;
-  }
-
-  const knownExistingNames = new Set(
-    [
-      currentDocument.name,
-      ...(existingFactoryNames ?? []),
-    ]
-      .map((name) => name.trim())
-      .filter((name) => name.length > 0),
-  );
-
-  return knownExistingNames.has(normalizedCreateFactoryName);
-}
-
-function toActivatedFactoryValue(document: CurrentFactoryDocument): FactoryValue {
-  const { version: _version, ...factoryValue } = document;
-  return factoryValue;
-}
-
 function resolveNamedFactoryAPIErrorCode(
   error: CurrentFactoryDefinitionError,
 ): NamedFactoryAPIErrorCode {
@@ -290,7 +151,9 @@ function resolveNamedFactoryAPIErrorCode(
     return normalizeNamedFactoryAPIErrorCode(errorBody.code);
   }
 
-  return normalizeNamedFactoryAPIErrorCodeFromCurrentFactoryDefinition(error.code);
+  return normalizeNamedFactoryAPIErrorCodeFromCurrentFactoryDefinition(
+    error.code,
+  );
 }
 
 function toNamedFactoryAPIErrorFromCurrentFactoryDefinition(
@@ -313,7 +176,24 @@ function toNamedFactoryAPIErrorFromCurrentFactoryDefinition(
     return new NamedFactoryAPIError(error.message, { code: "INTERNAL_ERROR" });
   }
 
-  return new NamedFactoryAPIError("Factory activation failed.", { code: "INTERNAL_ERROR" });
+  return new NamedFactoryAPIError("Factory activation failed.", {
+    code: "INTERNAL_ERROR",
+  });
+}
+
+function toNamedFactoryAPIErrorFromSessionFactory(
+  error: unknown,
+): NamedFactoryAPIError {
+  if (error instanceof SessionFactoryAPIError) {
+    return new NamedFactoryAPIError(error.message, {
+      code: error.code,
+      responseBody: error.responseBody,
+      status: error.status,
+      statusText: error.statusText,
+    });
+  }
+
+  return toNamedFactoryAPIErrorFromCurrentFactoryDefinition(error);
 }
 
 function normalizeNamedFactoryAPIErrorCodeFromCurrentFactoryDefinition(
